Add a business and a data access layer
Each time a command is run, a new session is created. A command may trigger
several connections to the database.

The expense package now contains:
* controller.go: the business layer
* datastore.go: the data access layer
* expense.go: the struct of the expense collection

The command calls the business layer, and the business layer uses the
datastore. The datastore makes a session.Copy() for each call to the database.