Refactoring on the session management
It was hard to make sure the session was well managed in the datastore. Also, it was necessary to test a little this management.

The database access is initialized when running a command. A main session is opened. Datastores will then be able to clone the main session to perform their changes.

Use the ExecuteInSession method to make sure that the session is opened and closed when running a query.