Refactoring in order to mock
After testing the mockery tool, it appears that it is not stable enough. The output is missing imports. The offical golang/mockgen is now used along with golang/gomock for the tests.

It was necessary to split Entities into Entities + DAO. The Entity holds the attributes of the collection, while the DAO is in charge of running the queries on the collection.